requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
google-generativeai>=0.3.0
Pillow>=9.5.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
pandas>=2.0.0
spacy>=3.6.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
gunicorn>=21.0.0